requirements.txt
Django==4.1.4
django-modeltranslation==0.18.7
environs==9.5.0
